fix: 충돌 해결 및 마이 페이지 폴더 구조 수정
